Moved telebot handler to telebot
