+++ rampart/browsing.py
@@ -3,4 +3,4 @@ from flask import Flask
 if __name__ == '__main__':
     app = Flask('rampart.browsing')
 
-    app.run('0.0.0.0', 9211, True, load_dotenv=False)
+    app.run('0.0.0.0', 9211, load_dotenv=False, use_reloader=False)
